Cleaned up code, handled some of the TODOs, fixed some of the warnings raised by the AppCode inspection, etc.

+++ JREngage/Classes/JRInfoBar.h
@@ -1,21 +1,21 @@
-/* * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * 
+/* * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *
  Copyright (c) 2010, Janrain, Inc.
-    
+
  All rights reserved.
-    
+
  Redistribution and use in source and binary forms, with or without modification,
  are permitted provided that the following conditions are met:
-    
+
  * Redistributions of source code must retain the above copyright notice, this
-     list of conditions and the following disclaimer. 
- * Redistributions in binary form must reproduce the above copyright notice, 
+     list of conditions and the following disclaimer.
+ * Redistributions in binary form must reproduce the above copyright notice,
      this list of conditions and the following disclaimer in the documentation and/or
-     other materials provided with the distribution. 
+     other materials provided with the distribution.
  * Neither the name of the Janrain, Inc. nor the names of its
      contributors may be used to endorse or promote products derived from this
      software without specific prior written permission.
-    
-    
+
+
  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
  ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
  WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
@@ -25,9 +25,9 @@
  LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
  ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
  (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE. 
-    
- File:   JRInfoBar.h 
+ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+ File:   JRInfoBar.h
  Author: Lilli Szafranski - lilli@janrain.com, lillialexis@gmail.com
  Date:   Tuesday, June 1, 2010
 * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * */
@@ -48,10 +48,10 @@ typedef enum
     UIView          *barBackground;
     UILabel         *poweredByLabel;
     UIButton        *infoButton;
-            
+
     BOOL hidesPoweredBy;
     BOOL iPad;
-    NSInteger y_origin_hidden;
+    CGFloat y_origin_hidden;
 }
 
 - (id)initWithFrame:(CGRect)frame andStyle:(JRInfoBarStyle)style;
